rename app package for clarity
